refactor common code into PhysicsBase
