+++ lib/network/modules/PhysicsWorker.js
@@ -1,29 +1,21 @@
-import BarnesHutSolver                      from './components/physics/BarnesHutSolver';
-import Repulsion                            from './components/physics/RepulsionSolver';
-import HierarchicalRepulsion                from './components/physics/HierarchicalRepulsionSolver';
-import SpringSolver                         from './components/physics/SpringSolver';
-import HierarchicalSpringSolver             from './components/physics/HierarchicalSpringSolver';
-import CentralGravitySolver                 from './components/physics/CentralGravitySolver';
-import ForceAtlas2BasedRepulsionSolver      from './components/physics/FA2BasedRepulsionSolver';
-import ForceAtlas2BasedCentralGravitySolver from './components/physics/FA2BasedCentralGravitySolver';
+import PhysicsBase                          from './PhysicsBase';
 
-class PhysicsWorker {
+class PhysicsWorker extends PhysicsBase {
   constructor(postMessage) {
+    super();
     this.body = {
       nodes: {},
       edges: {}
     };
-    this.physicsBody = {physicsNodeIndices:[], physicsEdgeIndices:[], forces: {}, velocities: {}};
     this.postMessage = postMessage;
-    this.options = {};
-    this.stabilized = false;
     this.previousStates = {};
-    this.positions = {};
-    this.timestep = 0.5;
     this.toRemove = {
       nodeIds: [],
       edgeIds: []
     };
+    this.physicsTimeout = null;
+    this.isWorker = true;
+    this.emit = (event, data) => {this.postMessage({type: 'emit', data: {event: event, data: data}})};
   }
 
   handleMessage(event) {
@@ -31,15 +23,10 @@ class PhysicsWorker {
     switch (msg.type) {
       case 'physicsTick':
         this.physicsTick();
+        this.sendPositions();
         break;
       case 'updatePositions':
-        let updatedNode = this.body.nodes[msg.data.id];
-        if (updatedNode) {
-          updatedNode.x = msg.data.x;
-          updatedNode.y = msg.data.y;
-          this.physicsBody.forces[updatedNode.id] = {x: 0, y: 0};
-          this.physicsBody.velocities[updatedNode.id] = {x: 0, y: 0};
-        }
+        this.receivePositions(msg.data);
         break;
       case 'updateProperties':
         this.updateProperties(msg.data);
@@ -48,71 +35,76 @@ class PhysicsWorker {
         this.addElements(msg.data);
         break;
       case 'removeElements':
-        // schedule removal of elements on the next physicsTick
-        // avoids having to defensively check every node read in each physics implementation
-        this.toRemove.nodeIds.push.apply(this.toRemove.nodeIds, msg.data.nodeIds);
-        this.toRemove.edgeIds.push.apply(this.toRemove.edgeIds, msg.data.edgeIds);
+        this.removeElements(msg.data);
+        break;
+      case 'stabilization':
+        this.stabilize(msg.data);
         break;
       case 'initPhysicsData':
+        console.debug('init physics data');
         this.initPhysicsData(msg.data);
         break;
       case 'options':
         this.options = msg.data;
         this.timestep = this.options.timestep;
-        this.init();
+        this.initPhysicsSolvers();
         break;
       default:
         console.warn('unknown message from PhysicsEngine', msg);
     }
   }
 
-  /**
-   * configure the engine.
-   */
-  init() {
-    var options;
-    if (this.options.solver === 'forceAtlas2Based') {
-      options = this.options.forceAtlas2Based;
-      this.nodesSolver = new ForceAtlas2BasedRepulsionSolver(this.body, this.physicsBody, options);
-      this.edgesSolver = new SpringSolver(this.body, this.physicsBody, options);
-      this.gravitySolver = new ForceAtlas2BasedCentralGravitySolver(this.body, this.physicsBody, options);
-    }
-    else if (this.options.solver === 'repulsion') {
-      options = this.options.repulsion;
-      this.nodesSolver = new Repulsion(this.body, this.physicsBody, options);
-      this.edgesSolver = new SpringSolver(this.body, this.physicsBody, options);
-      this.gravitySolver = new CentralGravitySolver(this.body, this.physicsBody, options);
-    }
-    else if (this.options.solver === 'hierarchicalRepulsion') {
-      options = this.options.hierarchicalRepulsion;
-      this.nodesSolver = new HierarchicalRepulsion(this.body, this.physicsBody, options);
-      this.edgesSolver = new HierarchicalSpringSolver(this.body, this.physicsBody, options);
-      this.gravitySolver = new CentralGravitySolver(this.body, this.physicsBody, options);
-    }
-    else { // barnesHut
-      options = this.options.barnesHut;
-      this.nodesSolver = new BarnesHutSolver(this.body, this.physicsBody, options);
-      this.edgesSolver = new SpringSolver(this.body, this.physicsBody, options);
-      this.gravitySolver = new CentralGravitySolver(this.body, this.physicsBody, options);
+//  physicsTick() {
+//    if(this.physicsTimeout) {
+//      // cancel any outstanding requests to prevent manipulation of data while iterating
+//      // and we're going to handle it next anyways.
+//      clearTimeout(this.physicsTimeout);
+//    }
+//    this.processRemovals();
+//    if (this.options.enabled) {
+//      this.calculateForces();
+//      this.moveNodes();
+//      // Handle the case where physics was enabled, data was removed
+//      // but this physics tick was longer than the timeout and during that delta
+//      // physics was disabled.
+//      this.processRemovals();
+//    }
+//    this.stabilizationIterations++;
+//  }
+
+  sendPositions() {
+    let nodeIndices = this.physicsBody.physicsNodeIndices;
+    let positions = {};
+    for (let i = 0; i < nodeIndices.length; i++) {
+      let nodeId = nodeIndices[i];
+      let node = this.body.nodes[nodeId];
+      positions[nodeId] = {x:node.x, y:node.y};
     }
 
-    this.modelOptions = options;
-  }
-
-  physicsTick() {
-    this.processRemovals();
-    this.calculateForces();
-    this.moveNodes();
-    for (let i = 0; i < this.toRemove.nodeIds.length; i++) {
-      delete this.positions[this.toRemove.nodeIds[i]];
-    }
     this.postMessage({
       type: 'positions',
       data: {
-        positions: this.positions,
+        positions: positions,
         stabilized: this.stabilized
       }
     });
+  }
+
+  receivePositions(data) {
+    let updatedNode = this.body.nodes[data.id];
+    if (updatedNode) {
+      updatedNode.x = data.x;
+      updatedNode.y = data.y;
+      this.physicsBody.forces[updatedNode.id] = {x: 0, y: 0};
+      this.physicsBody.velocities[updatedNode.id] = {x: 0, y: 0};
+    }
+  }
+
+  stabilize(data) {
+    this.stabilized = false;
+    this.targetIterations = data.targetIterations;
+    this.stabilizationIterations = 0;
+    setTimeout(() => this._stabilizationBatch(), 0);
   }
 
   updateProperties(data) {
@@ -157,10 +149,6 @@ class PhysicsWorker {
       if (replaceElements) {
         this.body.nodes[nodeId] = newNode;
       }
-      this.positions[nodeId] = {
-        x: newNode.x,
-        y: newNode.y
-      };
       this.physicsBody.forces[nodeId] = {x: 0, y: 0};
       // forces can be reset because they are recalculated. Velocities have to persist.
       if (this.physicsBody.velocities[nodeId] === undefined) {
@@ -182,6 +170,25 @@ class PhysicsWorker {
     }
   }
 
+  removeElements(data) {
+    // schedule removal of elements on the next physicsTick
+    // avoids having to defensively check every node read in each physics implementation
+    this.toRemove.nodeIds.push.apply(this.toRemove.nodeIds, data.nodeIds);
+    this.toRemove.edgeIds.push.apply(this.toRemove.edgeIds, data.edgeIds);
+    // Handle case where physics is disabled.
+    if(this.physicsTimeout) {
+      // don't schedule more than one physicsTick
+      clearTimeout(this.physicsTimeout);
+    }
+    this.physicsTimeout = setTimeout(()=> {
+      // if physics is still enabled, the next tick will handle removeElements
+      if (!this.options.enabled) {
+        this.physicsTimeout = null;
+        this.physicsTick();
+      }
+    }, 250);
+  }
+
   processRemovals() {
     while (this.toRemove.nodeIds.length > 0) {
       let nodeId = this.toRemove.nodeIds.pop();
@@ -191,7 +198,6 @@ class PhysicsWorker {
       }
       delete this.physicsBody.forces[nodeId];
       delete this.physicsBody.velocities[nodeId];
-      delete this.positions[nodeId];
       delete this.body.nodes[nodeId];
     }
     while (this.toRemove.edgeIds.length > 0) {
@@ -204,16 +210,10 @@ class PhysicsWorker {
     }
   }
 
-  /**
-   * Nodes and edges can have the physics toggles on or off. A collection of indices is created here so we can skip the check all the time.
-   *
-   * @private
-   */
   initPhysicsData(data) {
     this.physicsBody.forces = {};
     this.physicsBody.physicsNodeIndices = [];
     this.physicsBody.physicsEdgeIndices = [];
-    this.positions = {};
 
     this.body.nodes = data.nodes;
     this.body.edges = data.edges;
@@ -225,26 +225,6 @@ class PhysicsWorker {
         delete this.physicsBody.velocities[nodeId];
       }
     }
-  }
-
-  /**
-   * move the nodes one timestap and check if they are stabilized
-   * @returns {boolean}
-   */
-  moveNodes() {
-    var nodeIndices = this.physicsBody.physicsNodeIndices;
-    var maxVelocity = this.options.maxVelocity ? this.options.maxVelocity : 1e9;
-    var maxNodeVelocity = 0;
-
-    for (let i = 0; i < nodeIndices.length; i++) {
-      let nodeId = nodeIndices[i];
-      let nodeVelocity = this._performStep(nodeId, maxVelocity);
-      // stabilized is true if stabilized is true and velocity is smaller than vmin --> all nodes must be stabilized
-      maxNodeVelocity = Math.max(maxNodeVelocity,nodeVelocity);
-    }
-
-    // evaluating the stabilized and adaptiveTimestepEnabled conditions
-    this.stabilized = maxNodeVelocity < this.options.minVelocity;
   }
 
   /**
@@ -275,7 +255,6 @@ class PhysicsWorker {
       forces[nodeId].x = 0;
       velocities[nodeId].x = 0;
     }
-    this.positions[nodeId].x = node.x;
 
     if (node.options.fixed.y === false) {
       let dy   = this.modelOptions.damping * velocities[nodeId].y;    // damping force
@@ -288,19 +267,19 @@ class PhysicsWorker {
       forces[nodeId].y = 0;
       velocities[nodeId].y = 0;
     }
-    this.positions[nodeId].y = node.y;
 
     let totalVelocity = Math.sqrt(Math.pow(velocities[nodeId].x,2) + Math.pow(velocities[nodeId].y,2));
     return totalVelocity;
   }
 
-  /**
-   * calculate the forces for one physics iteration.
-   */
-  calculateForces() {
-    this.gravitySolver.solve();
-    this.nodesSolver.solve();
-    this.edgesSolver.solve();
+  _finalizeStabilization() {
+    this.sendPositions();
+    this.postMessage({
+      type: 'finalizeStabilization',
+      data: {
+        stabilizationIterations: this.stabilizationIterations
+      }
+    });
   }
 }
